Added a way to check if valid settings are present
Settings could be wrong when certain files are present. Without a
warning message it could be difficult to figure out what went wrong.

This commit adds a warning message if less/sass/style files are detected
but no valid compiler settings are detected.

Testing wise this is difficult/hacky as it would involve some sort of
output capturing. Otherwise I could return a string and print it, but
since this is an one time function that self destruct (which it does
not necessarily have to be) and a function that is not critical, it is
not included for now.